feat: add xtractor plugin, split image
The xtractor plugin depends on Essentia.
The image was split into two version:
* a small image (~100MB) without any plugins but the web API (for servers).
* a large image (~750MB) with many plugins (to be used to import/inspect the library).
This is because the image was already very large and adding essentia made it even bigger but when running the API no plugins other than the API are needed and it should be possible to run the API with minimal resource consumption on a e.g. a Raspberry Pi.